% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wiki_trends_and_history.R
\name{get_article_full_history_table}
\alias{get_article_full_history_table}
\title{Get all edits in a sorted table of wikipedien english for a given article name.}
\usage{
get_article_full_history_table(article_name)
}
\arguments{
\item{article_name}{Name of wikipedia article in string (i.e. Circadian clock)}
}
\value{
full edits history of the wikipedia article given as input.
}
\description{
Get all edits in a sorted table of wikipedien english for a given article name.
}
\examples{
get_article_full_history_table("Jeffrey C. Hall")
}
